Agent: Rename readme_utils to readme_dropper
