+++ monkey/infection_monkey/ransomware/readme_dropper.py
